Marginally dubious optimisation: treat memory and register imports as constants.  We're not trying to produce a generic program summary here, but are instead trying to determine which bits of synchronisation are relevant to a particular program execution.  As such, anything which isn't directly relevant to the synchronisation pattern can be discarded.  The physical contents of memory mostly fits into that category.
Of course, it is possible that memory contents controls what
synchronisation is needed.  In that case, we'll end up producing
something which is too pessimistic (i.e. performs synchronisation
unnecessarily), but still correct.  At this stage I'm willing to trade
a little bit of pessimism for a whole lot of simplicity.