Be a bit more aggressive in introducing free variables.  We can turn a load into a free if there are no there are no other loads which provably alias with it, not just if all other loads provably don't alias with it.
The reasoning here is moderately subtle.  Essentially, the only
downside of introducing free variables for loads is that you lose
track of where the thing was loaded from.  If A and B are loaded from
the same location, and we can prove that they're the same location,
then under many situations you can assume A == B.  If we replace the
loads with independent free variables then we can't make that
assumption any more, which means that we'll spend longer and generate
worse results.  However, if we *can't* prove that they're the same
location, we also can't make the assumption.  That means that if we
can't prove they're the same, there are no disadvantages to
introducing free variables for them, and doing so can make the
machines quite a bit simpler.