Rework the expression sorting phase so that we only sort as far as we need to:
-- Constants must go at the beginning
-- pointer-equal expressions must be adjacent to each other if the operation
   is commutative.

This is much faster, but makes things slightly less deterministic.
Fixing this up requires more calls to internIRExpr() in
eval_state_machine, and a whole lot of extra plumbing to know when to
do so.  Put it all in.

(In principle the nondeterminism could have broken us before this
change, we were just begin lucky).

Overall, this is a ~35% win in performance, which is worth having.