``More correct'' interpreter.  Make sure that we correctly handle sub-word loads of registers.
We still don't quite handle references to e.g. the ah, bh, ch, dh,
etc.  registers correctly, because if, say, you load rdx we won't
notice that a write to dh should have affected it, but apart from that
this is now sound.

(i.e. if you're just loading or storing at a reduced type in a
register then it's handled correctly, but if you're selecting a
sub-lane other than the first one then it isn't.  Hopefully that won't
matter too much.)