The write machine suitability constraint is a predict on machine states, P, such that for any state S which satisfies P:
-- Running the read machine in S leads to a non-crash
-- Running the write machine in S causes the write machine to reach a crash
   state, and running the read machine from that state also leads to
   a non-crash.

We already have a predict P' which is true of states where running the
read machine leads to a non-crash, so we can instead compute P'' such
that running the write machine from a P'' state causes the write
machine to crash in a P' state.  That's pretty much precisely what
writeMachineCrashConstraint does, so just use that.  This is *much*
faster than the current interpretation-based approach (multiple orders
of magnitude), although deriving the suitability constraint is usually
a relatively small fragment of total run time, so it's not that big a
win overall.