When building cross-product machines, we currently only consider the case where the probe machine issues at least one load before the store machine finishes.  This means that we don't handle store machines which look like this:
if (x) {
   foo = ...;
   bar = ...;
}

i.e. where there's some condition and the store machine effectively
doesn't run if it isn't true.  Editing out the case where the machine
doesn't run at all means that we lose the condition on x from the
cross-product machine, and if we've used it anywhere else that can
mean that we report bugs which don't really exist.

Basically, this patch makes us stop doing that, so that we now do
consider running the store machine to completion before starting the
load machine.  The write machine suitability constraint should prevent
this from introducing any new bugs, and the extra information should
remove a couple from consideration.