Further extend the phase 3 canonicaliser with two new passes:
1) Equality substitution: if we have a clause which says ``tmp1 == tmp2''
then we can replace all instances of tmp2 with tmp1, and doing so will
usually improve the canonical-ness of the summary.  If we have multiple
such equalities available then we decide which to use based on the number
of times tmp1 and tmp2 appear in the summary, prefering to remove whichever
register appears most often.  That's a pretty arbitrary heuristic, but
seems to work well for at least a couple of interesting examples.

2) Underspecified clause removal: if a register is used in precisely
one place, there's a good chance that it's not really telling us
anything. E.g. if we have ``x == y + z + w'' then we can treat that
clause as defining z.  If there are no other mentions of z anywhere in
the summary (machines and verification condition) then that's a
redundant definition and we can remove the clause from the
verification condition in the same way that we'd remove dead
assignments from the machines in the summary.