Extend the underspecification analysis with a kind of conditional underspecification, SSA-style analysis.
The idea is that if the verification condition V is a function of
boolean variables X, Y and Z then we should be able to do a case split
on any of them, X say, and re-express the condition as
(X && V(X=true)) || (!X && V(X=false)).  Once you've done that, you
can split Y into Y_{X=true} and Y_{X=false}, and likewise Z, in effect
treating all of our free variables as functions of X.  That then makes
all of the existing redundant clause and underspecified clause
canonicalisation passes much more powerful.

(You need to be careful that Y and Z are actually free, but,
fortunately, the existing targetRegisters set provides almost
precisely the right rules for doing so.)

There are a couple of ways of looking at this:

1) Most obviously, it captures certain conditional independence
   properties of the program.  It might by, for instance, that we're
   only interested in variable X and V is

   ((X == Y) || (X == Z)) && (Y != Z)

   If Y is true then that becomes

   ((X == true) || (X == Z)) && (true != Z)
   ->
   ((X == true) || (X == Z)) && (Z == false)
   ->
   (X == true) || (X == false)
   ->
   true.

   Likewise, if Y is false then it becomes

   ((X == false) || (X == Z)) && (false != Z)
   ->
   ((X == false) || (X == Z)) && (Z == true)
   ->
   (X == false) || (X == true)
   ->
   true

   V is therefore independent of X, conditional on Y, and so isn't
   really telling us anything.

   We can therefore re-express V as just

   (Y && true) || (!Y && true)
   ->
   Y || !Y
   ->
   true

   The case where we split on X first is somewhat more subtle, because
   as soon as you do the substitution you find that every clause is
   redundant (because it doesn't mention any target values), but
   produces essentially the same result.

2) It's also quite a lot like SSA.  The point is that the direct
   expression of the condition mixes up the values of Y for the
   different values of X, even though they're logically unrelated.  By
   splitting them up again we open up lots of other opportunities
   for optimisation.