Rework the way we find roots of store CFGs.  Previously, the only requirement was that every CFG node be reachable from at least one root.  That turns out to be inadequate, because some of the edges in the CFG aren't actually possible.
e.g. suppose you have a program like this:

l1: loc = a;
l2: for (i = 0; i < 1000000; i++)
l3:    nop;
l4: loc = b;

Now the CFG shows a simple path l1 -> l2 -> l4, so l1 and l4 will be
grouped together.  However, it's not actually possible to do that,
because you have to go round the loop 1000000 times.  The unroll won't
be able to cover that many instructions, though, so we'll in effect
only consider running the loop up to N_w times, which means that we'll
conclude that l4 can never run.  The only real fix to this is to make
l4 a root of the CFG as well.

We can still backtrack roots where its unambiguous to do so, though.
If we've selected X as a root, and Y is definitely executed
immediately before X, we can safely use Y as a root instead of X.
Because we know that prior to executing X we must have used the Y->X
edge, we know that if you can eliminate the Y->X edge then X is never
executable, so we're fine.