Whoops: if you get something like this:
a: STORE x -> loc1
b: LOAD y <- loc1
c: LOAD z <- loc1

Then realias will try to do forwarding a->b and b->c.  If it happens
to do the a->b one before the b->c one then when it comes to do the
second one b will have been transformed into a copy, and so so need to
support forwarding from copies as well as loads.