Make sure that we always drop the lock at the end of the patch, even if we encounter a ret instruction along the way.
This is arguably a bit of a hack.  The way we construct CFGs means
that rets should be no-ops except for adjusting RSP, becasue we
effectively inline cross-function stuff into the CFG and then rely on
the initial stack context check to make sure we're doing the right
thing, so we can just turn ret into lea 8(%rsp), %rsp.